Add logger (for console) (#98)
* Change run permissions on file

* Remove id from all the json data

* Add logger for the console

* Format logger

* Add termcolor

* Bump version

* Bump version

* Move version back

It's very interesting how consistent I am at times, I bumped the version
and committed the change without processing it, then did so again.
I definitely have a lot on my mind now, and it allows for this type of
thing to happen. I've found myself to make the same change twice, or
write something down a redundant amount. I often don't take many notes,
or at least it's not my main method of recalling information. Opening a
page of notes is definitely slow at scale, even firebase is faster.
It's very interesting tying to understand how I work, and very much so
how I've been able to make using the tools and languages I use almost
second nature. Often I'll write a script for something, put it in a
folder by name, put that folder in my Repos/ folder where I keep all my
many projects. Then a year or so later, I find the need for the same
script and once every so often, instead of recalling that I made a script
for it, I make I directory with the name that I think I should allot
this new project, this sometimes leads to an error from mkdir "cannot
create directory ‘<project-name>/’: File exists" and I realize I have
the exact thing I need and no memory if writing it. Rarely do I not have
it saved with git, I can tell when and with which key I used to sign the
commit, if I signed the commit for that given machine, and I usually can
find some clues as to what I was doing. Almost always however, I find in
depth documentation for why I needed such thing and how to use it. Only
a small fraction of these script end up on github, moreover, end up
being public. If you happen to be reading this commit message in the
future, send me an email at me@jr0.org and include the commit hash.

Cheers,
Jake Roggenbuck